+++ src/base/varpos.f90
@@ -405,16 +405,12 @@ endif
 ! or dilatable algorithm
 if (ivofmt.ge.0.or.idilat.gt.1) then
   call field_set_n_previous(icrom, 2)
-  if (iroext.gt.0) then
-    call field_set_n_previous(ibrom, 2)
-  endif
+  call field_set_n_previous(ibrom, 2)
   ! The density at the previous time step is required if
   ! we perform a hydrostatic pressure correction (icalhy=1)
 else if (iroext.gt.0.or.icalhy.eq.1.or.ipthrm.eq.1.or.ippmod(icompf).ge.0) then
   call field_set_n_previous(icrom, 1)
-  if (iroext.gt.0) then
-    call field_set_n_previous(ibrom, 1)
-  endif
+  call field_set_n_previous(ibrom, 1)
 endif
 ! Dans le cas d'une extrapolation de la viscosite totale
 if (iviext.gt.0) then